requirements.txt
torch~=2.2.0+cu118
torchvision~=0.17.0+cu118
opencv-python~=4.9.0.80
numpy~=1.26.4
matplotlib~=3.8.2
pillow~=10.2.0